Restructured/Refactored project
The project files were rearranged for greater clarity, and to service the refactor.

The refactor changed the client such that it is not accessing the entity framework directly, rather it interacts with an interface. An additional class which implements this interface has been constructed.

This should allow for greater extensibility and ease of use.