Added UI elements for import mapping
Working on #5.

A second page was added with the intention of showing what name the importer is using for a given property.

Similarly, a settings class was added with the intention of preserving any changes to the mapping between sessions.

These components are stable, but have yet to be connected.