The import is now working as it was prior to the refactor.
The BulkInsertCommand was broken in two ways: it was using an IRemsDbContext, not an IRemsDbFactory, and the ConvertNullableValue function was being passed null references, as during the refactor the datarow was switched from using its indexer to a Field<object>() method, which apparently contains an implicit conversion that the indexer does not use, meaning DBNull objects were now being returned as null references instead DBNull objects. This issue was further obscured by only occuring inside a LINQ select expression, meaning the error didn't actually occur until the enumerator was iterated over, hiding its location. This was not fun to debug, but the import now appears to be functional.