Swapped tabs and nodes in experiment detailer
Previously the treatments and plots were listed under the treeview, and the various charts were selected through a system of tabpages.

In order to bring the application more in line with classic REMS, these have been flipped, with the charts now available through the treeview, and individual treatments appearing as tabs on the right.

This is only the first step in reworking the detailer to be more like the original REMS.