Added handling for 'Design' table in the import
When a trait node is initialised it now checks if its parent is the 'Design' table - the enforcement of expected tables guarantees the existence of a design table appearing at some point in the import process.

If it is a child of the design table, a traitnode then switches out what the 'Add' function does, to include the data as a factor instead of a trait. The UI is updated accordingly.