Rewrote how the exporter queries soil data
Instead of each child model of soil having its own query, a single query looks for all the traits it expects to find and returns the results as a dictionary. This uses lets query and makes it much easier for the exporter to replace missing values with default ones from the template soil model.