Added a reworked version of SoilLayerData querying
The initial version of querying series data for charts performed a single query for each series, and returned them inside a custom SeriesData object. Additionally, there was a separate query for the mean data.

The new approach returns the data for each plot in a DataTable, with one column per plot/repetition, and an additional plot for the average. This requires far fewer queries (1 per trait/per treatment instead of 1 per plot/per treatment) and as a result should be much faster.

Additionally, the intent is each trait table is saved to a DataSet the first time it is called. This will further reduce the number of queries made, at the cost of additional memory.